+++ packages/nti-lib-scripts/tasks/build.js
@@ -27,7 +27,9 @@ call('node', [require.resolve('./check')]);
 call('node', [require.resolve('./test'), '--no-cache']);
 
 //Blank out lib
-fs.emptyDirSync(path.resolve(paths.path, 'lib'));
+if (process.env.NODE_ENV === 'production') {
+	fs.emptyDirSync(path.resolve(paths.path, 'lib'));
+}
 
 buildBundle()
 	.then(() => console.log(chalk.green('\nDone.\n\n')));
